Import same function in the index.js
